requirements.txt
flask
pdfplumber
python-pptx
